+++ app/Http/Controllers/Admin/AdminPhanHoiController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\Models\YeuThich;
+use App\Http\Controllers\Controller;
+use App\Models\AdminPhanHoi;
 use Illuminate\Http\Request;
 
-class YeuThichController extends Controller
+class AdminPhanHoiController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -34,7 +35,7 @@ class YeuThichController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(YeuThich $yeuThich)
+    public function show(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -42,7 +43,7 @@ class YeuThichController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(YeuThich $yeuThich)
+    public function edit(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -50,7 +51,7 @@ class YeuThichController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, YeuThich $yeuThich)
+    public function update(Request $request, AdminPhanHoi $adminPhanHoi)
     {
         //
     }
@@ -58,7 +59,7 @@ class YeuThichController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(YeuThich $yeuThich)
+    public function destroy(AdminPhanHoi $adminPhanHoi)
     {
         //
     }
