+++ app/Http/Controllers/Admin/BannerController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\BannerRequest;
 use App\Http\Requests\SuaBannerRequest;
 use App\Models\Banner;
