+++ app/Http/Controllers/Admin/BinhLuanController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\BinhLuan;
 use Illuminate\Http\Request;
