+++ app/Http/Controllers/Admin/DanhGiaController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\DanhGia;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 
 class DanhGiaController extends Controller
 {
