+++ app/Http/Controllers/Admin/DonHangController.php
@@ -1,18 +1,21 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\Models\KhuyenMaiSach;
+use App\Http\Controllers\Controller;
+use App\Models\DonHang;
 use Illuminate\Http\Request;
 
-class KhuyenMaiSachController extends Controller
+class DonHangController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        //
+        $listDonHang = DonHang::with(['sach', 'user', 'phuongThucThanhToan'])->orderByDesc('id')->get();
+
+        return view('admin.don-hang.index', compact('listDonHang'));
     }
 
     /**
@@ -34,7 +37,7 @@ class KhuyenMaiSachController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(KhuyenMaiSach $khuyenMaiSach)
+    public function show(DonHang $donHang)
     {
         //
     }
@@ -42,7 +45,7 @@ class KhuyenMaiSachController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(KhuyenMaiSach $khuyenMaiSach)
+    public function edit(DonHang $donHang)
     {
         //
     }
@@ -50,7 +53,7 @@ class KhuyenMaiSachController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, KhuyenMaiSach $khuyenMaiSach)
+    public function update(Request $request, DonHang $donHang)
     {
         //
     }
@@ -58,7 +61,7 @@ class KhuyenMaiSachController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(KhuyenMaiSach $khuyenMaiSach)
+    public function destroy(DonHang $donHang)
     {
         //
     }
