+++ app/Http/Controllers/Admin/MaGiamGiaController.php
@@ -1,20 +1,19 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
-use App\Models\DonHang;
+use App\Http\Controllers\Controller;
+use App\Models\MaGiamGia;
 use Illuminate\Http\Request;
 
-class DonHangController extends Controller
+class MaGiamGiaController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $listDonHang = DonHang::with(['sach', 'user', 'phuongThucThanhToan'])->orderByDesc('id')->get();
-       
-        return view('admin.don-hang.index', compact('listDonHang'));
+        //
     }
 
     /**
@@ -36,7 +35,7 @@ class DonHangController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(DonHang $donHang)
+    public function show(MaGiamGia $maGiamGia)
     {
         //
     }
@@ -44,7 +43,7 @@ class DonHangController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(DonHang $donHang)
+    public function edit(MaGiamGia $maGiamGia)
     {
         //
     }
@@ -52,7 +51,7 @@ class DonHangController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, DonHang $donHang)
+    public function update(Request $request, MaGiamGia $maGiamGia)
     {
         //
     }
@@ -60,7 +59,7 @@ class DonHangController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(DonHang $donHang)
+    public function destroy(MaGiamGia $maGiamGia)
     {
         //
     }
