+++ app/Providers/RouteServiceProvider.php
@@ -13,7 +13,7 @@ class RouteServiceProvider extends ServiceProvider
     /**
      * The path to your application's "home" route.
      *
-     * Typically, users are redirected here after authentication.
+     * Typically, user are redirected here after authentication.
      *
      * @var string
      */
