+++ database/seeders/DatabaseSeeder.php
@@ -17,7 +17,7 @@ class DatabaseSeeder extends Seeder
     {
 
         for ($i = 1; $i <= 10; $i++) {
-            DB::table('users')->insert([
+            DB::table('user')->insert([
                 'ten_doc_gia' => fake()->text(20),
                 'email' => fake()->email(),
                 'mat_khau' => fake()->password(6,15),
